+++ examples/run.sh
@@ -2,11 +2,9 @@
 #SBATCH --time=04:00:00
 #SBATCH --nodes=1
 #SBATCH --ntasks-per-node=16
-##SBATCH --partition=cs
 #SBATCH --partition=eng-instruction
+##SBATCH --partition=cs
 #SBATCH --mem=128GB
-
-time=/usr/bin/time
 
 # example to run TIPP3 with a given refpkg and a set of query reads
 bin=../run_tipp3.py
@@ -14,8 +12,8 @@ inpath=data/illumina.small.queries.fasta.gz
 t=16
 
 # NOTICE: supplement your own path of refpkg here
-refpkg=$HOME/Desktop/Research/phd_project/tipp3/tipp3-refpkg/
-#refpkg=$HOME/tallis/tipp3/tipp3-refpkg/
+#refpkg=$HOME/Desktop/Research/phd_project/tipp3/tipp3-refpkg/
+refpkg=$HOME/tallis/tipp3/tipp3-refpkg/
 #refpkg=../refpkg_scripts/custom_tipp_refpkg
 
 scenario=1
@@ -45,23 +43,30 @@ elif [[ $scenario == 3 ]]; then
         --alignment-method witch --placement-method pplacer-taxtastic \
         -t $t --keeptemp
 elif [[ $scenario == 4 ]]; then
-    # TIPP3-fast (BLAST+BSCAMPP) with .gz input type (fasta file)
-    inpath=data/nanopore.queries.fasta.gz
+    # TIPP3-fast for species detection, with a custom detection threshold B=0.3
     outdir=tipp3_scenario4
-    $bin abundance -i ${inpath} --reference-package ${refpkg} --outdir ${outdir} \
-        --alignment-method blast --placement-method bscampp \
-        -t $t
-elif [[ $scenario == 5 ]]; then
-    # TIPP3-fast (BLAST+BSCAMPP) with .gz input type (fastq file)
-    inpath=data/illumina.small.queries.fq.gz
-    outdir=tipp3_scenario5
-    $bin abundance -i ${inpath} --reference-package ${refpkg} --outdir ${outdir} \
-        --alignment-method blast --placement-method bscampp \
-        -t $t
-elif [[ $scenario == 6 ]]; then
-    # TIPP3-fast with species detection
-    outdir=tipp3_scenario6
+    detection_threshold=0.3
     $bin detection -i ${inpath} --reference-package ${refpkg} --outdir ${outdir} \
-        --alignment-method blast --placement-method bscampp \
-        -t $t
+        -t $t -B ${detection_threshold} --keeptemp
 fi
+#elif [[ $scenario == 4 ]]; then
+#    # TIPP3-fast (BLAST+BSCAMPP) with .gz input type (fasta file)
+#    inpath=data/nanopore.queries.fasta.gz
+#    outdir=tipp3_scenario4
+#    $bin abundance -i ${inpath} --reference-package ${refpkg} --outdir ${outdir} \
+#        --alignment-method blast --placement-method bscampp \
+#        -t $t
+#elif [[ $scenario == 5 ]]; then
+#    # TIPP3-fast (BLAST+BSCAMPP) with .gz input type (fastq file)
+#    inpath=data/illumina.small.queries.fq.gz
+#    outdir=tipp3_scenario5
+#    $bin abundance -i ${inpath} --reference-package ${refpkg} --outdir ${outdir} \
+#        --alignment-method blast --placement-method bscampp \
+#        -t $t
+#elif [[ $scenario == 6 ]]; then
+#    # TIPP3-fast with species detection
+#    outdir=tipp3_scenario6
+#    $bin detection -i ${inpath} --reference-package ${refpkg} --outdir ${outdir} \
+#        --alignment-method blast --placement-method bscampp \
+#        -t $t
+#fi