+++ tipp3/__init__.py
@@ -25,7 +25,7 @@ Updated @ 4.16.2025 by Chengze Shen
 
 Major changes to suit TIPP3 pipeline.
 '''
-__version__ = "0.3"
+__version__ = "0.4a"
 _INSTALL_PATH = __path__[0]
 
 __all__ = ['read_binning', 'read_alignment', 'read_placement',